Trait bdk_esplora::EsploraAsyncExt

pub trait EsploraAsyncExt {
    fn scan<'life0, 'life1, 'async_trait, K>(
        &'life0 self,
        local_chain: &'life1 BTreeMap<u32, BlockHash>,
        keychain_spks: BTreeMap<K, impl 'async_trait + IntoIterator<IntoIter = impl 'async_trait + Iterator<Item = (u32, Script)> + Send> + Send>,
        txids: impl 'async_trait + IntoIterator<IntoIter = impl 'async_trait + Iterator<Item = Txid> + Send> + Send,
        outpoints: impl 'async_trait + IntoIterator<IntoIter = impl 'async_trait + Iterator<Item = OutPoint> + Send> + Send,
        stop_gap: usize,
        parallel_requests: usize
    ) -> Pin<Box<dyn Future<Output = Result<LocalUpdate<K, ConfirmationTimeAnchor>, Error>> + Send + 'async_trait>>
    where
        K: 'async_trait + Ord + Clone + Send,
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;

    fn scan_without_keychain<'life0, 'life1, 'async_trait>(
        &'life0 self,
        local_chain: &'life1 BTreeMap<u32, BlockHash>,
        misc_spks: impl 'async_trait + IntoIterator<IntoIter = impl 'async_trait + Iterator<Item = Script> + Send> + Send,
        txids: impl 'async_trait + IntoIterator<IntoIter = impl 'async_trait + Iterator<Item = Txid> + Send> + Send,
        outpoints: impl 'async_trait + IntoIterator<IntoIter = impl 'async_trait + Iterator<Item = OutPoint> + Send> + Send,
        parallel_requests: usize
    ) -> Pin<Box<dyn Future<Output = Result<LocalUpdate<(), ConfirmationTimeAnchor>, Error>> + Send + 'async_trait>>
    where
        Self: Sync + 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
    { ... }
}

Trait to extend [esplora_client::AsyncClient] functionality.

This is the async version of EsploraExt. Refer to crate-level documentation for more.

Required Methods

fn scan<'life0, 'life1, 'async_trait, K>(
&'life0 self,
local_chain: &'life1 BTreeMap<u32, BlockHash>,
keychain_spks: BTreeMap<K, impl 'async_trait + IntoIterator<IntoIter = impl 'async_trait + Iterator<Item = (u32, Script)> + Send> + Send>,
txids: impl 'async_trait + IntoIterator<IntoIter = impl 'async_trait + Iterator<Item = Txid> + Send> + Send,
outpoints: impl 'async_trait + IntoIterator<IntoIter = impl 'async_trait + Iterator<Item = OutPoint> + Send> + Send,
stop_gap: usize,
parallel_requests: usize
) -> Pin<Box<dyn Future<Output = Result<LocalUpdate<K, ConfirmationTimeAnchor>, Error>> + Send + 'async_trait>>where
K: 'async_trait + Ord + Clone + Send,
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Scan the blockchain (via esplora) for the data specified and returns a [LocalUpdate<K, ConfirmationTimeAnchor>].

local_chain: the most recent block hashes present locally

keychain_spks: keychains that we want to scan transactions for

txids: transactions for which we want updated [ConfirmationTimeAnchor]s

outpoints: transactions associated with these outpoints (residing, spending) that we want to included in the update

The scan for each keychain stops after a gap of stop_gap script pubkeys with no associated transactions. parallel_requests specifies the max number of HTTP requests to make in parallel.

Provided Methods

fn scan_without_keychain<'life0, 'life1, 'async_trait>(
&'life0 self,
local_chain: &'life1 BTreeMap<u32, BlockHash>,
misc_spks: impl 'async_trait + IntoIterator<IntoIter = impl 'async_trait + Iterator<Item = Script> + Send> + Send,
txids: impl 'async_trait + IntoIterator<IntoIter = impl 'async_trait + Iterator<Item = Txid> + Send> + Send,
outpoints: impl 'async_trait + IntoIterator<IntoIter = impl 'async_trait + Iterator<Item = OutPoint> + Send> + Send,
parallel_requests: usize
) -> Pin<Box<dyn Future<Output = Result<LocalUpdate<(), ConfirmationTimeAnchor>, Error>> + Send + 'async_trait>>where
Self: Sync + 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Convenience method to call scan without requiring a keychain.