Trait bdk_esplora::EsploraExt

pub trait EsploraExt {
    fn full_scan<K: Ord + Clone>(
        &self,
        request: FullScanRequest<K>,
        stop_gap: usize,
        parallel_requests: usize
    ) -> Result<FullScanResult<K>, Error>;
    fn sync(
        &self,
        request: SyncRequest,
        parallel_requests: usize
    ) -> Result<SyncResult, Error>;
}

Trait to extend the functionality of [esplora_client::BlockingClient].

Refer to crate-level documentation for more.

Required Methods

fn full_scan<K: Ord + Clone>(
&self,
request: FullScanRequest<K>,
stop_gap: usize,
parallel_requests: usize
) -> Result<FullScanResult<K>, Error>

Scan keychain scripts for transactions against Esplora, returning an update that can be applied to the receiving structures.

local_tip: the previously seen tip from LocalChain::tip.

keychain_spks: keychains that we want to scan transactions for

The full scan for each keychain stops after a gap of stop_gap script pubkeys with no associated transactions. parallel_requests specifies the max number of HTTP requests to make in parallel.

Note

stop_gap is defined as “the maximum number of consecutive unused addresses”. For example, with a stop_gap of 3, full_scan will keep scanning until it encounters 3 consecutive script pubkeys with no associated transactions.

This follows the same approach as other Bitcoin-related software, such as Electrum, BTCPay Server, and Sparrow.

A stop_gap of 0 will be treated as a stop_gap of 1.

fn sync(
&self,
request: SyncRequest,
parallel_requests: usize
) -> Result<SyncResult, Error>

Sync a set of scripts with the blockchain (via an Esplora client) for the data specified and return a [TxGraph].

local_tip: the previously seen tip from LocalChain::tip.

misc_spks: scripts that we want to sync transactions for

txids: transactions for which we want updated [ConfirmationTimeHeightAnchor]s

outpoints: transactions associated with these outpoints (residing, spending) that we want to include in the update

If the scripts to sync are unknown, such as when restoring or importing a keychain that may include scripts that have been used, use full_scan with the keychain.