bdk_esplora
Trait EsploraAsyncExt

pub trait EsploraAsyncExt {
    // Required methods
    fn full_scan<'life0, 'async_trait, K, R>(
        &'life0 self,
        request: R,
        stop_gap: usize,
        parallel_requests: usize,
    ) -> Pin<Box<dyn Future<Output = Result<FullScanResponse<K>, Box<Error>>> + Send + 'async_trait>>
       where K: 'async_trait + Ord + Clone + Send,
             R: 'async_trait + Into<FullScanRequest<K>> + Send,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn sync<'life0, 'async_trait, I, R>(
        &'life0 self,
        request: R,
        parallel_requests: usize,
    ) -> Pin<Box<dyn Future<Output = Result<SyncResponse, Box<Error>>> + Send + 'async_trait>>
       where I: 'async_trait + Send,
             R: 'async_trait + Into<SyncRequest<I>> + Send,
             Self: 'async_trait,
             'life0: 'async_trait;
}

Trait to extend the functionality of [esplora_client::AsyncClient].

Refer to crate-level documentation for more.

Required Methods

fn full_scan<'life0, 'async_trait, K, R>( &'life0 self, request: R, stop_gap: usize, parallel_requests: usize, ) -> Pin<Box<dyn Future<Output = Result<FullScanResponse<K>, Box<Error>>> + Send + 'async_trait>>
where K: 'async_trait + Ord + Clone + Send, R: 'async_trait + Into<FullScanRequest<K>> + Send, Self: 'async_trait, 'life0: 'async_trait,

Scan keychain scripts for transactions against Esplora, returning an update that can be applied to the receiving structures.

request provides the data required to perform a script-pubkey-based full scan (see [FullScanRequest]). The full scan for each keychain (K) stops after a gap of stop_gap script pubkeys with no associated transactions. parallel_requests specifies the maximum number of HTTP requests to make in parallel.

Refer to crate-level docs for more.

fn sync<'life0, 'async_trait, I, R>( &'life0 self, request: R, parallel_requests: usize, ) -> Pin<Box<dyn Future<Output = Result<SyncResponse, Box<Error>>> + Send + 'async_trait>>
where I: 'async_trait + Send, R: 'async_trait + Into<SyncRequest<I>> + Send, Self: 'async_trait, 'life0: 'async_trait,

Sync a set of scripts, txids, and/or outpoints against Esplora.

request provides the data required to perform a script-pubkey-based sync (see [SyncRequest]). parallel_requests specifies the maximum number of HTTP requests to make in parallel.

Refer to crate-level docs for more.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.