Trait bdk_esplora::EsploraAsyncExt

pub trait EsploraAsyncExt {
    fn full_scan<'life0, 'async_trait, K>(
        &'life0 self,
        request: FullScanRequest<K>,
        stop_gap: usize,
        parallel_requests: usize
    ) -> Pin<Box<dyn Future<Output = Result<FullScanResult<K>, Box<Error>>> + Send + 'async_trait>>
    where
        K: 'async_trait + Ord + Clone + Send,
        Self: 'async_trait,
        'life0: 'async_trait;
    fn sync<'life0, 'async_trait>(
        &'life0 self,
        request: SyncRequest,
        parallel_requests: usize
    ) -> Pin<Box<dyn Future<Output = Result<SyncResult, Box<Error>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
}

Trait to extend the functionality of [esplora_client::AsyncClient].

Refer to crate-level documentation for more.

Required Methods

fn full_scan<'life0, 'async_trait, K>(
&'life0 self,
request: FullScanRequest<K>,
stop_gap: usize,
parallel_requests: usize
) -> Pin<Box<dyn Future<Output = Result<FullScanResult<K>, Box<Error>>> + Send + 'async_trait>>where
K: 'async_trait + Ord + Clone + Send,
Self: 'async_trait,
'life0: 'async_trait,

Scan keychain scripts for transactions against Esplora, returning an update that can be applied to the receiving structures.

request: struct with data required to perform a spk-based blockchain client full scan, see [FullScanRequest]

The full scan for each keychain stops after a gap of stop_gap script pubkeys with no associated transactions. parallel_requests specifies the max number of HTTP requests to make in parallel.

Note

stop_gap is defined as “the maximum number of consecutive unused addresses”. For example, with a stop_gap of 3, full_scan will keep scanning until it encounters 3 consecutive script pubkeys with no associated transactions.

This follows the same approach as other Bitcoin-related software, such as Electrum, BTCPay Server, and Sparrow.

A stop_gap of 0 will be treated as a stop_gap of 1.

fn sync<'life0, 'async_trait>(
&'life0 self,
request: SyncRequest,
parallel_requests: usize
) -> Pin<Box<dyn Future<Output = Result<SyncResult, Box<Error>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Sync a set of scripts with the blockchain (via an Esplora client) for the data specified and return a [TxGraph].

request: struct with data required to perform a spk-based blockchain client sync, see [SyncRequest]

If the scripts to sync are unknown, such as when restoring or importing a keychain that may include scripts that have been used, use full_scan with the keychain.