bdk_esplora
Trait EsploraExt

pub trait EsploraExt {
    // Required methods
    fn full_scan<K: Ord + Clone, R: Into<FullScanRequest<K>>>(
        &self,
        request: R,
        stop_gap: usize,
        parallel_requests: usize,
    ) -> Result<FullScanResponse<K>, Error>;
    fn sync<I: 'static, R: Into<SyncRequest<I>>>(
        &self,
        request: R,
        parallel_requests: usize,
    ) -> Result<SyncResponse, Error>;
}

Trait to extend the functionality of [esplora_client::BlockingClient].

Refer to crate-level documentation for more.

Required Methods

fn full_scan<K: Ord + Clone, R: Into<FullScanRequest<K>>>( &self, request: R, stop_gap: usize, parallel_requests: usize, ) -> Result<FullScanResponse<K>, Error>

Scan keychain scripts for transactions against Esplora, returning an update that can be applied to the receiving structures.

request provides the data required to perform a script-pubkey-based full scan (see FullScanRequest). The full scan for each keychain (K) stops after a gap of stop_gap script pubkeys with no associated transactions. parallel_requests specifies the maximum number of HTTP requests to make in parallel.

Refer to crate-level docs for more.

fn sync<I: 'static, R: Into<SyncRequest<I>>>( &self, request: R, parallel_requests: usize, ) -> Result<SyncResponse, Error>

Sync a set of scripts, txids, and/or outpoints against Esplora.

request provides the data required to perform a script-pubkey-based sync (see SyncRequest). parallel_requests specifies the maximum number of HTTP requests to make in parallel.

Refer to crate-level docs for more.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.