Trait bdk_esplora::EsploraExt

pub trait EsploraExt {
    fn scan<K: Ord + Clone>(
        &self,
        local_chain: &BTreeMap<u32, BlockHash>,
        keychain_spks: BTreeMap<K, impl IntoIterator<Item = (u32, Script)>>,
        txids: impl IntoIterator<Item = Txid>,
        outpoints: impl IntoIterator<Item = OutPoint>,
        stop_gap: usize,
        parallel_requests: usize
    ) -> Result<KeychainScan<K, ConfirmationTime>, Error>;

    fn scan_without_keychain(
        &self,
        local_chain: &BTreeMap<u32, BlockHash>,
        misc_spks: impl IntoIterator<Item = Script>,
        txids: impl IntoIterator<Item = Txid>,
        outpoints: impl IntoIterator<Item = OutPoint>,
        parallel_requests: usize
    ) -> Result<ChainGraph<ConfirmationTime>, Error> { ... }
}

Trait to extend [esplora_client::BlockingClient] functionality.

Refer to crate-level documentation for more.

Required Methods

fn scan<K: Ord + Clone>(
&self,
local_chain: &BTreeMap<u32, BlockHash>,
keychain_spks: BTreeMap<K, impl IntoIterator<Item = (u32, Script)>>,
txids: impl IntoIterator<Item = Txid>,
outpoints: impl IntoIterator<Item = OutPoint>,
stop_gap: usize,
parallel_requests: usize
) -> Result<KeychainScan<K, ConfirmationTime>, Error>

Scan the blockchain (via esplora) for the data specified and returns a KeychainScan.

local_chain: the most recent block hashes present locally

keychain_spks: keychains that we want to scan transactions for

txids: transactions for which we want updated ChainPositions

outpoints: transactions associated with these outpoints (residing, spending) that we want to included in the update

The scan for each keychain stops after a gap of stop_gap script pubkeys with no associated transactions. parallel_requests specifies the max number of HTTP requests to make in parallel.

Provided Methods

fn scan_without_keychain(
&self,
local_chain: &BTreeMap<u32, BlockHash>,
misc_spks: impl IntoIterator<Item = Script>,
txids: impl IntoIterator<Item = Txid>,
outpoints: impl IntoIterator<Item = OutPoint>,
parallel_requests: usize
) -> Result<ChainGraph<ConfirmationTime>, Error>

Convenience method to call scan without requiring a keychain.