Trait bdk_esplora::EsploraExt

pub trait EsploraExt {
    fn update_local_chain(
        &self,
        local_tip: CheckPoint,
        request_heights: impl IntoIterator<Item = u32>
    ) -> Result<Update, Error>;
    fn scan_txs_with_keychains<K: Ord + Clone>(
        &self,
        keychain_spks: BTreeMap<K, impl IntoIterator<Item = (u32, ScriptBuf)>>,
        txids: impl IntoIterator<Item = Txid>,
        outpoints: impl IntoIterator<Item = OutPoint>,
        stop_gap: usize,
        parallel_requests: usize
    ) -> Result<(TxGraph<ConfirmationTimeHeightAnchor>, BTreeMap<K, u32>), Error>;

    fn scan_txs(
        &self,
        misc_spks: impl IntoIterator<Item = ScriptBuf>,
        txids: impl IntoIterator<Item = Txid>,
        outpoints: impl IntoIterator<Item = OutPoint>,
        parallel_requests: usize
    ) -> Result<TxGraph<ConfirmationTimeHeightAnchor>, Error> { ... }
}

Trait to extend the functionality of [esplora_client::BlockingClient].

Refer to crate-level documentation for more.

Required Methods

fn update_local_chain(
&self,
local_tip: CheckPoint,
request_heights: impl IntoIterator<Item = u32>
) -> Result<Update, Error>

Prepare an LocalChain update with blocks fetched from Esplora.

prev_tip is the previous tip of LocalChain::tip.

get_heights is the block heights that we are interested in fetching from Esplora.

The result of this method can be applied to LocalChain::apply_update.

fn scan_txs_with_keychains<K: Ord + Clone>(
&self,
keychain_spks: BTreeMap<K, impl IntoIterator<Item = (u32, ScriptBuf)>>,
txids: impl IntoIterator<Item = Txid>,
outpoints: impl IntoIterator<Item = OutPoint>,
stop_gap: usize,
parallel_requests: usize
) -> Result<(TxGraph<ConfirmationTimeHeightAnchor>, BTreeMap<K, u32>), Error>

Scan Esplora for the data specified and return a TxGraph and a map of last active indices.

keychain_spks: keychains that we want to scan transactions for

txids: transactions for which we want updated ConfirmationTimeHeightAnchors

outpoints: transactions associated with these outpoints (residing, spending) that we want to include in the update

The scan for each keychain stops after a gap of stop_gap script pubkeys with no associated transactions. parallel_requests specifies the max number of HTTP requests to make in parallel.

Provided Methods

fn scan_txs(
&self,
misc_spks: impl IntoIterator<Item = ScriptBuf>,
txids: impl IntoIterator<Item = Txid>,
outpoints: impl IntoIterator<Item = OutPoint>,
parallel_requests: usize
) -> Result<TxGraph<ConfirmationTimeHeightAnchor>, Error>

Convenience method to call scan_txs_with_keychains without requiring a keychain.