Trait bdk_esplora::EsploraExt

pub trait EsploraExt {
    fn update_local_chain(
        &self,
        local_tip: CheckPoint,
        request_heights: impl IntoIterator<Item = u32>
    ) -> Result<Update, Box<Error>>;
    fn full_scan<K: Ord + Clone>(
        &self,
        keychain_spks: BTreeMap<K, impl IntoIterator<Item = (u32, ScriptBuf)>>,
        stop_gap: usize,
        parallel_requests: usize
    ) -> Result<(TxGraph<ConfirmationTimeHeightAnchor>, BTreeMap<K, u32>), Box<Error>>;
    fn sync(
        &self,
        misc_spks: impl IntoIterator<Item = ScriptBuf>,
        txids: impl IntoIterator<Item = Txid>,
        outpoints: impl IntoIterator<Item = OutPoint>,
        parallel_requests: usize
    ) -> Result<TxGraph<ConfirmationTimeHeightAnchor>, Box<Error>>;
}

Trait to extend the functionality of [esplora_client::BlockingClient].

Refer to crate-level documentation for more.

Required Methods

fn update_local_chain(
&self,
local_tip: CheckPoint,
request_heights: impl IntoIterator<Item = u32>
) -> Result<Update, Box<Error>>

Prepare a LocalChain update with blocks fetched from Esplora.

local_tip is the previous tip of LocalChain::tip.

request_heights is the block heights that we are interested in fetching from Esplora.

The result of this method can be applied to LocalChain::apply_update.

Consistency

The chain update returned is guaranteed to be consistent as long as there is not a large re-org during the call. The size of re-org we can tollerate is server dependent but will be at least 10.

fn full_scan<K: Ord + Clone>(
&self,
keychain_spks: BTreeMap<K, impl IntoIterator<Item = (u32, ScriptBuf)>>,
stop_gap: usize,
parallel_requests: usize
) -> Result<(TxGraph<ConfirmationTimeHeightAnchor>, BTreeMap<K, u32>), Box<Error>>

Full scan the keychain scripts specified with the blockchain (via an Esplora client) and returns a TxGraph and a map of last active indices.

keychain_spks: keychains that we want to scan transactions for

The full scan for each keychain stops after a gap of stop_gap script pubkeys with no associated transactions. parallel_requests specifies the max number of HTTP requests to make in parallel.

fn sync(
&self,
misc_spks: impl IntoIterator<Item = ScriptBuf>,
txids: impl IntoIterator<Item = Txid>,
outpoints: impl IntoIterator<Item = OutPoint>,
parallel_requests: usize
) -> Result<TxGraph<ConfirmationTimeHeightAnchor>, Box<Error>>

Sync a set of scripts with the blockchain (via an Esplora client) for the data specified and return a TxGraph.

misc_spks: scripts that we want to sync transactions for

txids: transactions for which we want updated ConfirmationTimeHeightAnchors

outpoints: transactions associated with these outpoints (residing, spending) that we want to include in the update

If the scripts to sync are unknown, such as when restoring or importing a keychain that may include scripts that have been used, use full_scan with the keychain.